+++ Changelog.md
@@ -13,6 +13,8 @@
   when databases have named clauses. The error "Error: unable to graft this
   clause: no clause named ..." should no more be raised in response to a
   Require Import.
+- New `coq.strategy.*` to `set` and `get` the unfolding priority of constants
+  followed by the term comparison algorithm Coq uses at ype checking time.
 
 ### HOAS
 - Fix handling of default case in `match`, now Coq's `if _ then _ else _`